email field and keyboard types
