+++ lib/widgets/genericTextButton.dart
@@ -1,0 +1,32 @@
+import 'package:flutter/material.dart';
+
+class GenericTextButton extends StatelessWidget {
+  final String text;
+  final VoidCallback onPressed;
+
+  GenericTextButton({
+    required this.text,
+    required this.onPressed,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      onTap: onPressed,
+      child: Container(
+        decoration: BoxDecoration(
+          borderRadius: BorderRadius.circular(15),
+          color: const Color.fromARGB(255, 6, 30, 82),
+        ),
+        height: 50,
+        width: 330,
+        child: Center(
+          child: Text(
+            text,
+            style: TextStyle(color: Colors.white, fontSize: 18),
+          ),
+        ),
+      ),
+    );
+  }
+}
